added monthly horrorscopes

+++ src/main/java/com/horror_scope/demo/emails/EmailService.java
@@ -3,6 +3,8 @@ package com.horror_scope.demo.emails;
 import com.horror_scope.demo.exception.Exceptions;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class EmailService{
 
@@ -14,6 +16,10 @@ public class EmailService{
         this.emailDAO = emailDAO;
     }
 
+    public Optional<Email> getEmail(String email) {
+        return emailDAO.selectEmailByEmail(email);
+    }
+    
     public void addEmail (String email, String zodiac){
 
         if (emailValidator.isValid(email)){
@@ -25,6 +31,11 @@ public class EmailService{
     }
 
     public void deleteEmail (String email){
+        Optional<Email> checkEmail = emailDAO.selectEmailByEmail(email);
+        
+        if(checkEmail.isEmpty()){
+            throw new Exceptions("Email not found");
+        }
         emailDAO.deleteEmail(email);
     }
